+++ lib/src/linked_in_profile.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 import 'package:flutter/material.dart';
+import 'package:linkedin_login/src/constants.dart';
 import 'package:linkedin_login/src/linked_in_auth_response_wrapper.dart';
 import 'package:linkedin_login/src/linked_in_user_model.dart';
 import 'package:linkedin_login/src/linked_in_authorization_webview.dart';
@@ -15,6 +16,7 @@ class LinkedInUserWidget extends StatefulWidget {
   final String clientId, clientSecret;
   final PreferredSizeWidget appBar;
   final bool destroySession;
+  final List<String> projection;
 
   /// Client state parameter needs to be unique range of characters - random one
   LinkedInUserWidget({
@@ -25,7 +27,21 @@ class LinkedInUserWidget extends StatefulWidget {
     this.catchError,
     this.destroySession = false,
     this.appBar,
-  });
+    this.projection = const [
+      ProjectionParameters.id,
+      ProjectionParameters.localizedFirstName,
+      ProjectionParameters.localizedLastName,
+      ProjectionParameters.firstName,
+      ProjectionParameters.lastName,
+    ],
+  })
+      : assert(onGetUserProfile != null),
+        assert(redirectUrl != null),
+        assert(clientId != null),
+        assert(clientSecret != null),
+        assert(destroySession != null),
+        assert(appBar != null),
+        assert(projection != null);
 
   @override
   State createState() => _LinkedInUserWidgetState();
@@ -34,12 +50,22 @@ class LinkedInUserWidget extends StatefulWidget {
 /// Class [_LinkedInUserWidgetState] is handling changes after user is singed in
 /// which will have as result user profile on the end
 class _LinkedInUserWidgetState extends State<LinkedInUserWidget> {
-  final urlLinkedInUserProfile = 'https://api.linkedin.com/v2/me';
+  String urlLinkedInUserProfile =
+      'https://api.linkedin.com/v2/me';
   final urlLinkedInEmailAddress =
       'https://api.linkedin.com/v2/emailAddress?q=members&projection=(elements*(handle~))';
 
   @override
-  Widget build(BuildContext context) => LinkedInAuthorization(
+  void initState() {
+    super.initState();
+
+    String projection  = 'projection=(${widget.projection.join(",")})';
+    urlLinkedInUserProfile = '$urlLinkedInUserProfile?$projection';
+  }
+
+  @override
+  Widget build(BuildContext context) =>
+      LinkedInAuthorization(
         destroySession: widget.destroySession,
         redirectUrl: widget.redirectUrl,
         clientSecret: widget.clientSecret,
@@ -52,7 +78,7 @@ class _LinkedInUserWidgetState extends State<LinkedInUserWidget> {
               headers: {
                 HttpHeaders.acceptHeader: 'application/json',
                 HttpHeaders.authorizationHeader:
-                    'Bearer ${result.accessToken.accessToken}'
+                'Bearer ${result.accessToken.accessToken}'
               },
             ).then((basicProfile) {
               get(
@@ -60,12 +86,12 @@ class _LinkedInUserWidgetState extends State<LinkedInUserWidget> {
                 headers: {
                   HttpHeaders.acceptHeader: 'application/json',
                   HttpHeaders.authorizationHeader:
-                      'Bearer ${result.accessToken.accessToken}'
+                  'Bearer ${result.accessToken.accessToken}'
                 },
               ).then((emailProfile) {
                 // Get basic user profile
                 final LinkedInUserModel linkedInUser =
-                    LinkedInUserModel.fromJson(json.decode(basicProfile.body));
+                LinkedInUserModel.fromJson(json.decode(basicProfile.body));
                 // Get email for current user profile
                 linkedInUser.email = LinkedInProfileEmail.fromJson(
                   json.decode(emailProfile.body),